Factor out Java and JUnit4-related code to separate modules #20
