+++ repomate_junit4/__init__.py
@@ -1,1 +1,3 @@
 from repomate_junit4.__version import __version__
+
+SECTION = "junit4"
